+++ package.json
@@ -16,6 +16,7 @@
     "core-js": "^3.23.4",
     "firebase": "^9.9.0",
     "next": "12.2.2",
+    "next-themes": "^0.2.0",
     "react": "18.2.0",
     "react-dom": "18.2.0"
   },
